@@ -1,9 +1,29 @@
 <?php
 
- use Faker\Generator as Faker;
+namespace Database\Factories;
 
- $factory->define(App\Categoria::class, function (Faker $faker) {
-    return [
-        'nome' => $faker->unique()->text(25),
-    ];
-});
+use App\Models\Categoria;
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+class CategoriaFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Categoria::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'nome' => $this->faker->unique()->text(25),
+        ];
+    }
+}
+
